move the user tests into the handlers directory
