+++ test/handlers/user/password.js
@@ -6,15 +6,15 @@ var Code = require('code'),
     after = lab.after,
     it = lab.test,
     expect = Code.expect,
-    redisSessions = require('../../adapters/redis-sessions');
+    redisSessions = require('../../../adapters/redis-sessions');
 
 var server, cookieCrumb,
-    fakeuser = require('../fixtures/users').fakeusercouch,
-    fakeChangePass = require('../fixtures/users').fakeuserChangePassword;
+    fakeuser = require('../../fixtures/users').fakeusercouch,
+    fakeChangePass = require('../../fixtures/users').fakeuserChangePassword;
 
 // prepare the server
 before(function (done) {
-  require('../mocks/server')(function (obj) {
+  require('../../mocks/server')(function (obj) {
     server = obj;
     done();
   });
